requirements.txt
certifi==2025.10.5
Flask==3.1.2
flask-cors==6.0.1
numpy==2.3.5
pymongo==3.12.0
python-dotenv==1.2.2
requests==2.32.5
